Extract SbtMavenRepository from MavenRepository
Remove sbtAttrsStub from MavenRepository
Remove extraAttributes handling from PomParser
