+++ modules/core/jvm/src/test/scala/coursier/maven/PomParserTests.scala
@@ -1,19 +1,19 @@
 package coursier
 
-import coursier.core.{Classifier, Configuration}
+import coursier.core.{Classifier, Configuration, Info, Project}
 import coursier.core.compatibility._
 import coursier.util.Traverse.TraverseOps
-import coursier.maven.MavenRepository
-import coursier.maven.Pom
+import coursier.maven.PomParser
 import utest._
-import coursier.core.Info
-import coursier.core.Info.License
 
 object PomParserTests extends TestSuite {
 
+  private def parseRawPomSax(str: String): Either[String, Project] =
+    coursier.core.compatibility.xmlParseSax(str, new PomParser).project
+
   val tests = Tests {
     test("scm field is optional") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |    <modelVersion>4.0.0</modelVersion>
@@ -28,7 +28,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("all fields in scm is optional") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |    <modelVersion>4.0.0</modelVersion>
@@ -48,7 +48,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("can parse scm info") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |    <modelVersion>4.0.0</modelVersion>
@@ -71,7 +71,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("properties are parsed") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |    <modelVersion>4.0.0</modelVersion>
@@ -91,7 +91,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("licenses are optional") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |  <modelVersion>4.0.0</modelVersion>
@@ -107,7 +107,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("licenses with just name and url") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |  <modelVersion>4.0.0</modelVersion>
@@ -136,7 +136,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("licenses with just name and url (binary compat test)") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |  <modelVersion>4.0.0</modelVersion>
@@ -160,7 +160,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("multiple licenses with just name and url") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |  <modelVersion>4.0.0</modelVersion>
@@ -199,7 +199,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("license with maven distribution and comments") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |  <modelVersion>4.0.0</modelVersion>
@@ -230,7 +230,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("license with maven distribution and comments (binary compat test)") {
-      val success = MavenRepository.parseRawPomSax(
+      val success = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |  <modelVersion>4.0.0</modelVersion>
@@ -256,7 +256,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("'/' and '\\' are invalid in groupId") {
-      val failure = MavenRepository.parseRawPomSax(
+      val failure = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |    <modelVersion>4.0.0</modelVersion>
@@ -271,7 +271,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("'/' and '\\' are invalid in artifactId") {
-      val failure = MavenRepository.parseRawPomSax(
+      val failure = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |    <modelVersion>4.0.0</modelVersion>
@@ -286,7 +286,7 @@ object PomParserTests extends TestSuite {
     }
 
     test("'/' and '\\' are invalid in version") {
-      val failure = MavenRepository.parseRawPomSax(
+      val failure = parseRawPomSax(
         """
           |<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
           |    <modelVersion>4.0.0</modelVersion>
