+++ modules/tests/shared/src/test/scala/coursier/test/CentralTests.scala
@@ -6,6 +6,7 @@ import utest._
 import scala.async.Async.{async, await}
 import coursier.core.{Classifier, Configuration, Extension, Type}
 import coursier.graph.{Conflict, ModuleTree}
+import coursier.maven.MavenRepositoryBase
 import coursier.test.compatibility._
 import coursier.util.{Artifact, Print, Tree}
 
@@ -15,15 +16,14 @@ object CentralTests extends CentralTests
 
 abstract class CentralTests extends TestSuite {
 
-  def central             = Repositories.central
-  private def centralBase = central.root
+  def central: MavenRepositoryBase = Repositories.central
+  private def centralBase          = central.root
 
   private final def isActualCentral = centralBase == Repositories.central.root
 
-  private lazy val runner = new TestRunner(repositories = Seq(central))
+  protected lazy val runner = new TestRunner(repositories = Seq(central))
 
   def tests = Tests {
-
     test("logback") {
       async {
         val dep = dep"ch.qos.logback:logback-classic:1.1.3"
@@ -641,53 +641,6 @@ abstract class CentralTests extends TestSuite {
         assert(pomOpt.forall(hasSha1))
         assert(pomOpt.forall(hasMd5))
         assert(pomOpt.forall(hasSig))
-      }
-    }
-
-    test("sbtPlugin") {
-      test("versionRange") {
-        val mod = mod"org.ensime:sbt-ensime;scalaVersion=2.10;sbtVersion=0.13"
-        val ver = "1.12.+"
-
-        test {
-          // doesn't work via proxies, which don't list all the upstream available versions
-          if (isActualCentral)
-            runner.resolutionCheck(mod, ver)
-          else
-            Future.successful(())
-        }
-      }
-
-      test("diamond") {
-        // sbt-plugin-example-diamond is a diamond graph of sbt plugins.
-        // Diamond depends on left and right which both depend on bottom.
-        //             sbt-plugin-example-diamond
-        //                        / \
-        // sbt-plugin-example-left   sbt-plugin-example-right
-        //                        \ /
-        //             sbt-plugin-example-bottom
-        // Depending on the version of sbt-plugin-example-diamond, different patterns
-        // are tested:
-        // - Some plugins are only published to the deprecated Maven path, some to the new
-        // - There may be some conflict resolution to perform on sbt-plugin-example-bottom,
-        //   mixing old and new Maven paths.
-        val diamond =
-          mod"ch.epfl.scala:sbt-plugin-example-diamond;scalaVersion=2.12;sbtVersion=1.0"
-
-        // only deprecated Maven paths
-        test("0.1.0") - runner.resolutionCheck(diamond, "0.1.0")
-
-        // diamond and left on the new Maven path
-        test("0.2.0") - runner.resolutionCheck(diamond, "0.2.0")
-
-        // conflict resolution on bottom between new and deprecated Maven paths
-        test("0.3.0") - runner.resolutionCheck(diamond, "0.3.0")
-
-        // bottom on new Maven paht, but not righ
-        test("0.4.0") - runner.resolutionCheck(diamond, "0.4.0")
-
-        // only new Maven paths with conflict resolution on bottom
-        test("0.5.0") - runner.resolutionCheck(diamond, "0.5.0")
       }
     }
